Widen bin_exp parameters to int64_t to recover hot-path performance
The narrowing of `to_decimal`'s `raw_exp`/`bin_exp` from int64_t to
int32_t (introduced when fixing the MSVC C4244 warnings on calls into
`compute_dec_exp` / `compute_exp_shift`) silenced the warnings at the
cost of measurable throughput. Each `static_cast<uint64_t>(bin_exp)` in
the hot path then required a sign-extension (`sxtw` on ARM64,
`movsxd` on x86-64) that was free when bin_exp was already 64-bit, and
`bin_exp + traits::exp_offset` used as an array index needed the same
extension at the indexing site.

Measured on Apple M1 Max, clang -O3 -march=native, three runs each
against the float_write_benchmark:

  Config             main    PR (narrow)    This commit
  float fast         1233    1211 (-1.8%)   1242 (+0.7%)
  float size-opt      898     862 (-4.0%)    891 (-0.8%)
  double fast        2740    2716 (-0.9%)   2792 (+1.9%)
  double size-opt    1602    1572 (-1.9%)   1602 (0.0%)
  std float (ctrl)    364     359            365
  std double (ctrl)   605     605            607

Throughput in MB/s. The std::to_chars columns confirm run-to-run noise
was small. The 4% size-opt regression on float is fully recovered.

The fix moves the narrowing from the carrier (the local) back to the
narrow points (the function returns), preserving the warning-free
build:

  - `compute_dec_exp(int64_t bin_exp) -> int32_t` with an explicit
    `static_cast<int32_t>` at the return statement.
  - `compute_exp_shift(int64_t bin_exp, int32_t dec_exp) -> int32_t`
    with an explicit `static_cast<int32_t>` at the return statement.
  - `to_decimal(UInt bin_sig, int64_t raw_exp, ...)` and the
    `int64_t bin_exp = raw_exp - traits::exp_offset;` local restored
    to int64_t.

Each MSVC warning the original PR addressed remains silenced:

  - C4244 `compute_dec_exp(bin_exp)` argument conversion is gone
    because the parameter is now int64_t.
  - C4244 `compute_exp_shift(bin_exp, ...)` argument conversion is
    gone for the same reason.
  - C4244 inside `compute_dec_exp` / `compute_exp_shift` returns are
    silenced by the explicit `static_cast<int32_t>`.
  - C4244 on `int dec_exp = umul128_hi64(...)` retains the PR's
    `static_cast<int32_t>`.
  - The signed-to-unsigned conversion `static_cast<uint64_t>(bin_exp)`
    feeding `umul128_hi64` is preserved verbatim; with int64_t bin_exp
    it is now codegen-free (a reinterpret) instead of a sxtw.
  - C4456 shadow renames (`irregular_dec_exp`, `float_extra_shift`)
    and all other casts in `write` are unchanged.

A subtle pitfall worth recording: an earlier draft of this fix tried
to return `uint8_t` from `compute_exp_shift`, matching the
pre-existing-PR signature, on the theory that a shift count fits in a
byte. This crashed `float_write_test` immediately. For `bin_exp = 972`
(max double normal), `bin_exp + pow10_bin_exp + 1` evaluates to -1.
The original code masked this by relying on a two-stage modular wrap
(unsigned char return, then `unsigned char shift = ... + extra_shift`,
which produced 5). The PR replaced both stages with full int32_t
carrying, which is the algorithmically correct change. Returning
`uint8_t` from the function alone, while the caller stored in
`int32_t shift`, broke the wrap and yielded an out-of-range shift
amount. Keep `int32_t` as the return type.

float_write_test: 99939 doubles + 99615 floats round-trip clean.

+++ include/glaze/util/zmij.hpp
@@ -323,12 +323,12 @@ namespace glz::zmij
          return ZMIJ_USE_INT128 ? umul128_hi64(x, div10_sig64) : x / 10;
       }
 
-      constexpr auto compute_dec_exp(int32_t bin_exp, bool regular = true) noexcept -> int32_t
+      constexpr auto compute_dec_exp(int64_t bin_exp, bool regular = true) noexcept -> int32_t
       {
          assert(bin_exp >= -1334 && bin_exp <= 2620);
          constexpr int log10_3_over_4_sig = 131'072;
          constexpr int log10_2_sig = 315'653, log10_2_exp = 20;
-         return (bin_exp * log10_2_sig - !regular * log10_3_over_4_sig) >> log10_2_exp;
+         return static_cast<int32_t>((bin_exp * log10_2_sig - !regular * log10_3_over_4_sig) >> log10_2_exp);
       }
 
       template <typename Float>
@@ -444,12 +444,12 @@ namespace glz::zmij
          }
       };
 
-      constexpr ZMIJ_INLINE auto compute_exp_shift(int32_t bin_exp, int32_t dec_exp) noexcept -> int32_t
+      constexpr ZMIJ_INLINE auto compute_exp_shift(int64_t bin_exp, int32_t dec_exp) noexcept -> int32_t
       {
          assert(dec_exp >= -350 && dec_exp <= 350);
          constexpr int log2_pow10_sig = 217'707, log2_pow10_exp = 16;
          int pow10_bin_exp = -dec_exp * log2_pow10_sig >> log2_pow10_exp;
-         return bin_exp + pow10_bin_exp + 1;
+         return static_cast<int32_t>(bin_exp + pow10_bin_exp + 1);
       }
 
       template <bool OptSize>
@@ -838,10 +838,10 @@ namespace glz::zmij
       };
 
       template <typename Float, typename UInt, bool OptSize>
-      ZMIJ_INLINE auto to_decimal(UInt bin_sig, int32_t raw_exp, bool regular, const constants<OptSize>& c) noexcept -> to_decimal_result
+      ZMIJ_INLINE auto to_decimal(UInt bin_sig, int64_t raw_exp, bool regular, const constants<OptSize>& c) noexcept -> to_decimal_result
       {
          using traits = float_traits<Float>;
-         int32_t bin_exp = raw_exp - traits::exp_offset;
+         int64_t bin_exp = raw_exp - traits::exp_offset;
          constexpr int num_bits = std::numeric_limits<UInt>::digits;
 
          constexpr uint64_t log10_2_sig = 78'913;
